fix: redundant call

+++ android-tv/app/src/main/java/studio/kahn/iris/tv/ui/screens/CollectionScreen.kt
@@ -381,7 +381,7 @@ private suspend fun doGrabPack(
         }.getOrNull()
     } ?: return
     if (autoPlay) {
-        onPlay(res.infohash, res.fileIdx.toInt())
+        onPlay(res.infohash, res.fileIdx)
     }
 }
